Change initial cancel time for agents, (see wiki); fix cancelAll*Orders methods  in MatchingEngine; add simulation state information to title bar; fixed factor calculation for HFT

+++ src/edu/virginia/jinsup/HFTPoisson.java
@@ -10,7 +10,7 @@ public class HFTPoisson extends PoissonAgent {
   @Override
   void makeOrder() {
     double factor =
-      getBestBuyPrice() / (getBestBuyPrice() + getBestSellPrice());
+      (double) getBestBuyPrice() / (getBestBuyPrice() + getBestSellPrice());
     boolean willBuy = true;
 
     // if shares own 20 cancel all buys and P(buy) = 0%
